PathTracker

PathTracker is a data structure to track for a given AbstractHomotopy $H(x,t)$ a solution $x$ from $t₁ > 0$ to $0$, i.e., $H(x,t₁) = 0$ and $x'$ with $H(x',0) = 0$ is returned. This is done by following an implicitly defined path $x(t)$ using Tracker. In contrast to Tracker this uses an endgame to handle diverging paths and singular solutions.

Constructor and Options

HomotopyContinuation2.PathTracker — Type

PathTracker(tracker::Tracker; options = PathTrackerOptions())
PathTracker(H::AbstractHomotopy; options = PathTrackerOptions())

A PathTracker combines a Tracker with an endgame. That is, while a Tracker assumes that the solution path is non-singular and convergent, the endgame allows to handle singular endpoints as well as diverging paths. To compute singular solutions the Cauchy endgame used, for divering paths a strategy based on the valuation of local Puiseux series expansion of the path is used. See [BT20] for a detailed description. By convention, a PathTracker always tracks from $t=1$ to $t = 0$. See PathTrackerOptions for the possible options.

HomotopyContinuation2.PathTrackerOptions — Type

PathTrackerOptions(; options...)

Options controlling the behaviour of a PathTracker.

Options

endgame_start::Float64 = 0.1: The point t in time where the endgame starts.

Endgame parameters

These parameters control the behaviour during the endgame. See [BT20] for details.

min_cond_eg::Float64 = 1e6: The minimal condition number for which an endgame strategy is applied.

at_infinity_check::Bool = true: Whether divering paths should be truncated.

zero_is_at_infinity::Bool = false: Whether paths going to a solution where at least one coordinates is zero should also be considered diverging.

val_finite_tol::Float64 = 1e-2: Tolerance on the valuation which has to be satisfied before the Cauchy endgame is started.

max_winding_number::Int = 20: The maximal winding number which is attempted in the Cauchy endgame.

val_at_infinity_tol::Float64 = 1e-3: Tolerance on the valuation which has to be satisfied before a path is considered to diverge / go to infinity.

Tracking

HomotopyContinuation2.track — Method

track(path_tracker::PathTracker, x::AbstractVector, t::Real = 1.0;
      path_number = nothing, debug = false)

Track the given start solution x from t towards 0 using the given path_tracker. Returns a PathResult.

track(path_tracker::PathTracker, r::PathResult, t::Real = 1.0;
      path_number = nothing, debug = false)

Track solution(r) from t towards 0 using the given path_tracker.

PathResult

HomotopyContinuation2.PathResult — Type

PathResult{V<:AbstractVector}

A PathResult is the result of tracking of a path with track using a PathTracker.

Fields

General solution information:

return_code: See the list of return codes below.

solution::V: The solution vector.

t::Float64: The value of t at which solution was computed. Note that if return_code is :at_infinity, then t is the value when this was decided.

accuracy::Float64: An estimate the (relative) accuracy of the computed solution.

residual::Float64: The infinity norm of H(solution,t).

condition_jacobian::Float64: This is the condition number of the Jacobian at the solution. A high condition number indicates a singular solution or a solution on a positive dimensional component.

winding_number:Union{Nothing, Int}: The computed winding number. This is a lower bound on the multiplicity of the solution. It is $nothing$ if the Cauchy endgame was not used.

extended_precision::Bool: Indicate whether extended precision is necessary to achieve the accuracy of the solution.

path_number::Union{Nothing,Int}: The number of the path (optional).

start_solution::Union{Nothing,V}: The start solution of the path (optional).

Performance information:

accepted_steps::Int: The number of accepted steps during the path tracking.

rejected_steps::Int: The number of rejected steps during the path tracking.

extended_precision_used::Bool: Indicates whether extended precision was necessary to track the path.

Additional path and solution informations

valuation::Vector{Float64}: An approximation of the valuation of the Puiseux series expansion of $x(t)$.

last_path_point::Tuple{V,Float64}: The last pair $(x,t)$ before the solution was computed. If the solution was computed with the Cauchy endgame, then the pair $(x,t)$ can be used to rerun the endgame.

Return codes

Possible return codes are:

:success: The PathTracker obtained a solution.

:at_infinity: The PathTracker stopped the tracking of the path since it determined that that path is diverging towards infinity.

:at_zero: The PathTracker stopped the tracking of the path since it determined that that path has a solution where at least one coordinate is 0. This only happens if the option zero_is_at_infinity is true.

:excess_solution: For the solution of the system, the system had to be modified which introduced artificial solutions and this solution is one of them.

various return codes indicating termination of the tracking

HomotopyContinuation2.solution — Method

solution(r::PathResult)

Get the solution of the path.

HomotopyContinuation2.is_success — Method

is_success(r::PathResult)

Checks whether the path is successfull.

HomotopyContinuation2.is_at_infinity — Method

is_at_infinity(r::PathResult)

Checks whether the path goes to infinity.

HomotopyContinuation2.is_excess_solution — Method

is_excess_solution(r::PathResult)

Checks whether the path is successfull.

HomotopyContinuation2.is_failed — Method

is_failed(r::PathResult)

Checks whether the path failed.

HomotopyContinuation2.is_finite — Method

is_finite(r::PathResult)

Checks whether the path result is finite.

HomotopyContinuation2.is_singular — Method

is_singular(r::PathResult; tol::Float64 = 1e10)

Checks whether the path result r is singular. This is true if the winding number is larger than 1 or if the condition number of the Jacobian is larger than tol.

HomotopyContinuation2.is_nonsingular — Method

is_nonsingular(r::PathResult; tol::Float64 = 1e10)

Checks whether the path result is non-singular. This is true if it is not singular.

HomotopyContinuation2.is_real — Method

is_real(r::PathResult; tol::Float64 = 1e-6)

We consider a result as real if the infinity-norm of the imaginary part of the solution is at most tol.

HomotopyContinuation2.accuracy — Method

accuracy(r::PathResult)

Get the accuracy of the solution. This is an estimate of the (relative) distance to the true solution.

HomotopyContinuation2.residual — Method

residual(r::PathResult)

Get the residual of the solution.

HomotopyContinuation2.steps — Method

steps(r::PathResult)

Total number of steps the path tracker performed.

HomotopyContinuation2.accepted_steps — Method

accepted_steps(r::PathResult)

Total number of steps the path tracker accepted.

HomotopyContinuation2.rejected_steps — Method

rejected_steps(r::PathResult)

Total number of steps the path tracker rejected.

HomotopyContinuation2.winding_number — Method

winding_number(r::PathResult)

Get the winding number of the solution of the path. Returns nothing if it wasn't computed.

HomotopyContinuation2.path_number — Method

path_number(r::PathResult)

Get the number of the path. Returns nothing if it wasn't provided.

HomotopyContinuation2.start_solution — Method

start_solution(r::PathResult)

Get the start solution of the path.

HomotopyContinuation2.multiplicity — Method

multiplicity(r::PathResult)

Get the multiplicity of the solution of the path. Returns nothing if it wasn't computed.

HomotopyContinuation2.last_path_point — Method

last_path_point(r::PathResult)

Returns a tuple (x,t) containing the last zero of H(x, t) before the Cauchy endgame was used. Returns nothing if the endgame strategy was not invoked.

HomotopyContinuation2.valuation — Method

valuation(r::PathResult)

Get the computed valuation of the path.

BT20Breiding, P. and Timme, S. "Tropical Endgame", In preparation (2020)

BT20Breiding, P. and Timme, S. "Tropical Endgame", In preparation (2020)