Start Systems

Total Degree

HomotopyContinuation2.total_degree — Function

total_degree(F::System; parameters = nothing, gamma = cis(2π * rand()))
total_degree(F::AbstractSystem; gamma = cis(2π * rand()))

Construct a total degree homotopy $H$ using the 'γ-trick' such that $H(x,0) = F(x)$ and $H(x,1)$ is the generated start system. Returns a StraightLineHomotopy and the start solutions as an interator.

HomotopyContinuation2.total_degree_start_solutions — Function

total_degree_start_solutions(degrees)

Returns an iterator of the start solutions of the system

\[\begin{array}{rl} z_1^{d_1} - 1 &= 0 \\ z_2^{d_2} - 1 &= 0 \\ &\vdots \\ z_n^{d_n} - 1 &= 0 \\ \end{array}\]

where $d_i$ is degrees[i].

Example

julia> iter = total_degree_start_solutions([2, 2])
4 total degree start solutions for degrees [2, 2]

julia> collect(iter)
4-element Array{Array{Complex{Float64},1},1}:
 [1.0 + 0.0im, 1.0 + 0.0im]
 [-1.0 + 1.2246467991473532e-16im, 1.0 + 0.0im]
 [1.0 + 0.0im, -1.0 + 1.2246467991473532e-16im]
 [-1.0 + 1.2246467991473532e-16im, -1.0 + 1.2246467991473532e-16im]

Mixed Volume (Polyhedral Homotopy)

HomotopyContinuation2.polyhedral — Function

polyhedral(F::Union{System, AbstractSystem};
    only_non_zero = false,
    path_tracker_options = PathTrackerOptions(),
    tracker_options = TrackerOptions())

Solve the system F in two steps: first solve a generic system derived from the support of F using a polyhedral homotopy as proposed in [HS95], then perform a coefficient-parameter homotopy towards F. This returns a PolyhedralTracker and an iterator to compute the start solutions.

If only_non_zero is true, then only the solutions with non-zero coordinates are computed. In this case the number of paths to track is equal to the mixed volume of the Newton polytopes of F.

If only_non_zero is false, then all isolated solutions of F are computed. In this case the number of paths to track is equal to the mixed volume of the convex hulls of $supp(F_i) ∪ \{0\}$ where $supp(F_i)$ is the support of $F_i$. See also [LW96].

function polyhedral(
    support::AbstractVector{<:AbstractMatrix},
    coefficientss::AbstractVector{<:AbstractVector{<:Number}};
    kwargs...,
)

It is also possible to provide directly the support and coefficients of the system F to be solved.

Example

We consider a system f which has in total 6 isolated solutions, but only 3 where all coordinates are non-zero.

@var x y
f = System([2y + 3 * y^2 - x * y^3, x + 4 * x^2 - 2 * x^3 * y])
tracker, starts = polyhedral(f; only_non_zero = false)
# length(starts) == 8
count(is_success, track.(tracker, starts)) # 6

tracker, starts = polyhedral(f; only_non_zero = true)
# length(starts) == 3
count(is_success, track.(tracker, starts)) # 3

HomotopyContinuation2.PolyhedralTracker — Type

PolyhedralTracker <: AbstractTracker

This tracker realises the two step approach of the polyhedral homotopy. See also [polyhedral].

HomotopyContinuation2.PolyhedralStartSolutionsIterator — Type

PolyhedralStartSolutionsIterator

An iterator providing start solutions for the polyhedral homotopy.

HS95Birkett Huber and Bernd Sturmfels. “A Polyhedral Method for Solving Sparse Polynomial Systems.” Mathematics of Computation, vol. 64, no. 212, 1995, pp. 1541–1555

LW96T.Y. Li and Xiaoshen Wang. "The BKK root count in C^n". Math. Comput. 65, 216 (October 1996), 1477–1484.